Added creation of Fields to TwoPlayer board
During the creation of the fields, if the field should start with a piece on it, the piece is found and assigned to the field. This makes use of the Discriminator to find the correct piece. I'm currently unsure of, if this is possible, as entity framework might get mad that i myself have set the discriminator property to a value, where entity framework itself would add the same exact value, but at the point of SaveChanges.